% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panelSummary.R
\name{panelSummary}
\alias{panelSummary}
\title{Summarize markers in panels.}
\usage{
panelSummary(panelInfo, cluster = TRUE, plotImage = TRUE)
}
\arguments{
\item{panelInfo}{A data frame returned by the collectData function. It should
contain all the information outputted by the preprocessing.batch function.}

\item{cluster}{True or False. Used to indicate if the markers and panels
should be clustered in the plot.}

\item{plotImage}{True or False. Used to indicate if a plot summarizing
markers in panels should be produced.}
}
\value{
A dataframe describing what markers are in each panel.
}
\description{
A function that summarizes markers in cytometry panels.
}
\examples{
fn=system.file("extdata","",package="MetaCyto")
fn=list.files(fn,pattern="processed_sample",full.names=TRUE)
panel_info=collectData(fn,longform=FALSE)
pdf("Panel_summary.pdf",width=30,height=20)
PS=panelSummary(panelInfo = panel_info,
                cluster=FALSE)
dev.off()
}
